Support first level domain website items

+++ src/models/cache.py
@@ -4,6 +4,7 @@ from typing import Optional, Any
 from pydantic import BaseModel, validate_arguments
 
 from src import console
+from src.models.exceptions import NotLoggedInError
 from src.models.ssdb_database import SsdbDatabase
 from src.models.wikimedia.wikipedia.templates.wikipedia_page_reference import (
     WikipediaPageReference,
@@ -18,6 +19,8 @@ class Cache(BaseModel):
     @validate_arguments
     def add_page(self, wikipedia_page: Any, wcdqid: str):
         logger.debug("add_page: Running")
+        if self.ssdb is None:
+            raise NotLoggedInError()
         if wikipedia_page.md5hash is not None and wcdqid is not None:
             # if type(reference.wikicitations_qid) is not str:
             #     raise ValueError(f"{reference.wikicitations_qid} is not of type str")
@@ -29,11 +32,26 @@ class Cache(BaseModel):
     @validate_arguments
     def add_reference(self, reference: WikipediaPageReference, wcdqid: str):
         logger.debug("add_reference: Running")
+        if self.ssdb is None:
+            raise NotLoggedInError()
         if reference.md5hash is not None and wcdqid is not None:
             # if type(reference.wikicitations_qid) is not str:
             #     raise ValueError(f"{reference.wikicitations_qid} is not of type str")
             logger.debug(f"Trying to set the value: {wcdqid}")
             return self.ssdb.set_value(key=reference.md5hash, value=wcdqid)
+        else:
+            raise ValueError("did not get what we need")
+
+    @validate_arguments
+    def add_website(self, reference: WikipediaPageReference, wcdqid: str):
+        logger.debug("add_website: Running")
+        if self.ssdb is None:
+            raise NotLoggedInError()
+        if reference.first_level_domain_of_url_hash is not None and wcdqid is not None:
+            logger.debug(f"Trying to set the value: {wcdqid}")
+            return self.ssdb.set_value(
+                key=reference.first_level_domain_of_url_hash, value=wcdqid
+            )
         else:
             raise ValueError("did not get what we need")
 
@@ -44,13 +62,15 @@ class Cache(BaseModel):
         wikipedia_page: Any,  # WikipediaPage
     ) -> Optional[str]:
         """We get binary from SSDB so we decode it"""
+        if self.ssdb is None:
+            raise NotLoggedInError()
         if wikipedia_page.md5hash is not None:
             # https://stackoverflow.com/questions/55365543/
             response = self.ssdb.get_value(key=wikipedia_page.md5hash)
             if response is None:
                 return None
             else:
-                return response.decode("UTF-8")
+                return str(response.decode("UTF-8"))
         else:
             raise ValueError("md5hash was None")
 
@@ -59,15 +79,35 @@ class Cache(BaseModel):
         self, reference: WikipediaPageReference
     ) -> Optional[str]:
         """We get binary from SSDB so we decode it"""
+        if self.ssdb is None:
+            raise NotLoggedInError()
         if reference.md5hash is not None:
             # https://stackoverflow.com/questions/55365543/
             response = self.ssdb.get_value(key=reference.md5hash)
             if response is None:
                 return None
             else:
-                return response.decode("UTF-8")
+                return str(response.decode("UTF-8"))
         else:
             raise ValueError("md5hash was None")
+
+    @validate_arguments
+    def check_website_and_get_wikicitations_qid(
+        self, reference: WikipediaPageReference
+    ) -> Optional[str]:
+        """We get binary from SSDB so we decode it"""
+        if self.ssdb is None:
+            raise NotLoggedInError()
+        if reference.first_level_domain_of_url_hash is not None:
+            # https://stackoverflow.com/questions/55365543/
+            response = self.ssdb.get_value(key=reference.first_level_domain_of_url_hash)
+            if response is None:
+                return None
+            else:
+                return str(response.decode("UTF-8"))
+        else:
+            # Not all references have urls so we fail silently
+            return None
 
     @validate_arguments
     def connect(self, host: str = "127.0.0.1", port: int = 8888):
@@ -80,7 +120,10 @@ class Cache(BaseModel):
 
     @validate_arguments
     def delete_key(self, key: str):
-        return self.ssdb.delete(key=key)
+        if self.ssdb is not None:
+            return self.ssdb.delete(key=key)
+        else:
+            raise ValueError("self.ssdb was None")
 
     def flush_database(self):
         result = self.ssdb.flush_database()
